% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descartar.R
\name{descartar}
\alias{descartar}
\title{Função de descarte de anos ou meses que apresentam falhas.}
\usage{
descartar(dados, falhas = 5, modo = "anual")
}
\arguments{
\item{dados}{Corresponde à data frame obtida após o uso da função
\code{organizar}.}

\item{falhas}{Valor numérico correspondente à porcentagem de falhas
permitidas no mês/ano.}

\item{modo}{Define se o modo de descarte é \code{anual} ou \code{mensal}.}
}
\description{
Descarta anos ou meses que apresentarem porcentagem de falhas
  superiores às especificadas pelo usuário.
}
\details{
Existem dois modos de descartes realizados: \code{anual} ou
  \code{mensal}. Serão descartados os anos/meses que possuirem valores acima
  do indicado pela variável \code{falhas}. O valor padrão para \code{falhas}
  é de 5\%.
}
\examples{
# Descartando anos com falhas iguais ou superiores a 5\%:
descartar(fluviopolis, falhas = 5, modo = "anual")

# Descartando meses com falhas iguais ou superiores a 10\%:
descartar(fluviopolis, falhas = 10, modo = "mensal")

}
