% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distprob.R
\name{distprob}
\alias{distprob}
\title{Função para ajuste de distribuições de probabilidades e
estimação de parâmetros.}
\usage{
distprob(valores, dist, prob)
}
\arguments{
\item{valores}{Um vetor numérico com os valores a serem ajustados.}

\item{dist}{Um vetor de caracteres com os nomes das distribuições de
probabilidade a serem utilizadas.}

\item{prob}{Um valor numérico, correspondente à probabilidade a qual o
parâmetro (vazão ou precipitação) será estimado.}
}
\description{
Função para ajuste de distribuições de probabilidades e
estimação de parâmetros.
}
