% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluviopolis.R
\name{fluviopolis}
\alias{fluviopolis}
\title{Dados da estação de Fluviópolis}
\description{
Dados das vazões (m³/s) do rio Iguaçu na estação de Fluviópolis,
  localizada no município de São Mateus do Sul - PR. Os dados são do período
  de 1963 a 2010.
}
