% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidrowebr.R
\docType{package}
\name{hidrowebr}
\alias{hidrowebr}
\alias{hidrowebr-package}
\title{hidrowebr: Uma ferramenta para análise estátistica de séries hidrológicas.}
\description{
As funções do hidrowebr são divididas em três categorias: pré-processamento,
processamento e análise.
}
\section{Pré-processamento}{

As funções de pré-processamento organizam os dados brutos

\code{\link{importar}}
\code{\link{organizar}}
\code{\link{descartar}}
}

\section{Processamento}{

As funções de processamento
}

\section{Análise}{

As funções de análise
\code{\link{distprob}}
}

