% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hidrograma}
\alias{plot_hidrograma}
\title{Funções de plotagem de hidrogramas.}
\usage{
plot_hidrograma(valores, datas, estacoes)
}
\arguments{
\item{valores}{Vetor numérico com as leituras a serem plotadas.}

\item{datas}{Vetor com as datas das leituras.}

\item{estacoes}{Vetor com as estações correspondentes a cada uma das
leituras.}
}
\description{
Funções de plotagem de hidrogramas.
}
\details{
O hidrograma gerado separa as diferentes estações por cores. Os
  gráficos são feitos com auxílio do pacote \code{\link[ggplot2]{ggplot2-package}}.
}
