% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qperm.R
\name{qperm}
\alias{qperm}
\title{Função para cálculo de vazões de permanência}
\usage{
qperm(cperm, freq)
}
\arguments{
\item{cperm}{Data frame obtida pela função de \code{cperm}.}

\item{freq}{Vetor numérico com as frequências para as vazões de permanência.}
}
\description{
Calcula as vazões de permanência obtida após a função de
  \code{\link{cperm}}.
}
\details{
As vazões de permanência são calculadas como a média do valor
  imediatamente superior e imediatamente inferior à frequência especificada.
}
\examples{
# Cálculo da curva de permanência:
cPerm <- cperm(fluviopolis)

# Cálculo das vazões de permanência:
qperm(cPerm, 95)

}
