% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qref.R
\name{qref}
\alias{qref}
\title{Função para cálculo de vazões de referência}
\usage{
qref(cperm, ref)
}
\arguments{
\item{cperm}{Data frame obtida pela função de \code{cperm}.}

\item{ref}{Vetor numérico com as frequências para as vazões de referência.}
}
\description{
Calcula as vazões de referência obtida após a função de
  \code{\link{cperm}}.
}
\details{
As vazões de referência são calculadas como a média do valor
  imediatamente superior e imediatamente inferior à frequência especificada.
}
