% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumoDist.R
\name{resumoDist}
\alias{resumoDist}
\title{Resumo dos resultados de \code{distprob}}
\usage{
resumoDist(distProb)
}
\arguments{
\item{distProb}{Data frame obtida pela função \code{\link{distprob}}.}
}
\description{
Resume os resultados da função \code{\link{distprob}},
  retornando apenas a melhor distribuição (aquela com menor valor no teste de
  Kolmogorov-Smirnov) e o valor estimado.
}
\examples{
# Cálculo de vazões máximas:
qmax <- maxAnuais(fluviopolis, "Q")

# Estimativa da vazão máxima para o período de retorno de 10 anos com a
# distribuição log-normal:
distProb <- distprob(qmax, "Maxima", dist = c("norm", "lnorm"), prob = 0.1)

#Resumo
resumoDist(distProb)

}
