% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descartar.R
\name{descartar}
\alias{descartar}
\title{Função para descartar anos ou meses que apresentam certa porcentagem de
falhas.}
\usage{
descartar(dados, falhas = 5, modo = "anual")
}
\arguments{
\item{dados}{Corresponde à data frame obtida após o uso da função
\code{organizar}.}

\item{falhas}{Valor numérico correspondente à porcentagem de falhas
permitidas no mês/ano.}

\item{modo}{Define se o modo de descarte é \code{anual} ou \code{mensal}.}
}
\description{
Função para descartar anos ou meses que apresentam certa porcentagem de
falhas.
}
\details{
Existem dois modos de descartes realizados: \code{anual} ou
  \code{mensal}. Serão descartados os anos/meses que possuirem valores acima
  do indicado pela variável \code{falhas}. O valor padrão para \code{falhas}
  é de 5\%.
}
