% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar.R
\name{importar}
\alias{importar}
\title{Função de importação dos dados de arquivos \code{txt} e \code{csv}.}
\usage{
importar(arq)
}
\arguments{
\item{arq}{Variável do tipo \code{chr}. Corresponde ao nome ou caminho do
arquivo a ser importado.}
}
\description{
Função de importação dos dados de arquivos \code{txt} e \code{csv}.
}
\details{
Por meio dessa função serão importados os dados tais quais estão
  dispostos no arquivo original do portal Hidroweb.
}
