% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anuais.R
\name{maxAnuais}
\alias{maxAnuais}
\title{Funções para determinar os valores máximos, mínimos ou médios anuais.}
\usage{
maxAnuais(valores, datas, estacoes, ano_hidro = "01-01")
}
\arguments{
\item{valores}{Vetor numérico com os valores de vazão ou precipitação.}

\item{datas}{Vetor de datas correspondentes aos dados.}

\item{estacoes}{Vetor com as correspondentes estações de cada um dos
elementos dos vetores anteriores.}

\item{ano_hidro}{Variável do tipo \code{chr} que contém o dia e mês a ser
utilizado como início do ano hidrológico. Tem como padrão o dia primeiro de
janeiro ("01-01").}
}
\description{
Funções para determinar os valores máximos, mínimos ou médios anuais.
}
\details{
No argumento \code{ano_hidro} deve-se repassar uma variável do tipo
  \code{chr} com o dia e o mês a ser utilizado para iniciar o ano
  hidrológico. Dessa maneira, a data ("15-02") corresponderia ao dia 15 de
  fevereiro, por exemplo. Caso o valor para o ano hidrológico seja uma data
  inválida, a função utilizará o dia "01-01" para determinar os valores
  anuais.
}
