% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar.R
\name{organizar}
\alias{organizar}
\title{Função de organização dos dados}
\usage{
organizar(dados, formato_data = c("Dia", "Mes", "Ano"))
}
\arguments{
\item{dados}{Os dados brutos obtidos após a função de \code{\link{importar}}.}

\item{formato_data}{Formatação das datas no arquivo.}
}
\description{
Organiza os dados em colunas de Estação, Data e
  Vazão/Precipitação.
}
\details{
Os dados brutos são organizados em uma data frame com: colunas de
  \code{Est}, que representa o código da estação; \code{Data}, que é a data
  da leitura; e \code{Q} ou \code{P}, valores de vazão (m³/s) ou precipitação
  (mm), respectivamente.

  O valor padrão para a variável \code{formato_data} é \code{Dia}, \code{Mes}
  e \code{Ano}. Deve-se ordernar esses três componentes conforme a formatação
  que estiver presente no arquivo do portal Hidroweb. A partir da versão
  2.0.0, o Hidroweb tem como padrão as datas na forma dd/mm/YYYY.

  A função também exclui os valores repetidos. A preferência é para aqueles
  que possuem nível de consistência 2 (nível consistido).
}
