% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plotagem de boxplots}
\usage{
plot_boxplot(dados, col_valores)
}
\arguments{
\item{dados}{Data frame com a série hidrológica a ser utilizada.}

\item{col_valores}{O nome da coluna com os valores a serem utilizados.}
}
\description{
Função de plotagem de boxplots.
}
\details{
O boxplot gerado separa os dados por estação. Os gráficos são feitos
  com auxílio do pacote \code{\link[ggplot2]{ggplot2-package}}.
}
\examples{
# Boxplot da estação de Fluviópolis:
plot_boxplot(fluviopolis, col_valores = "Q")

}
