% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Funções de plotagem de boxplots.}
\usage{
plot_boxplot(valores, estacoes, tipo = "Q")
}
\arguments{
\item{valores}{Vetor numérico com as leituras a serem plotadas.}

\item{estacoes}{Vetor com as estações correspondentes a cada uma das
leituras.}

\item{tipo}{Variável que corresponde ao tipo de leitura, as opções são:
\code{Q} ou \code{P}, para vazões ou precipitações, respectivamente.}

\item{datas}{Vetor com as datas das leituras.}
}
\description{
Funções de plotagem de boxplots.
}
\details{
O boxplot gerado separa os dados por estação. Os gráficos são feitos
  com auxílio do pacote \code{\link[ggplot2]{ggplot2-package}}.
}
