% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hidrograma}
\alias{plot_hidrograma}
\title{Função de plotagem de hidrogramas}
\usage{
plot_hidrograma(dados, col_valores)
}
\arguments{
\item{dados}{Data frame com a série hidrológica a ser utilizada.}

\item{col_valores}{O nome da coluna com os valores a serem utilizados.}
}
\description{
Função de plotagem de hidrogramas
}
\details{
O hidrograma gerado separa as diferentes estações por cores. Os
  gráficos são feitos com auxílio do pacote
  \code{\link[ggplot2]{ggplot2-package}}.
}
\examples{
# Hidrograma da estação de Fluviópolis:
plot_hidrograma(fluviopolis, col_valores = "Q")

}
