% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histograma}
\alias{plot_histograma}
\title{Funções de plotagem de histogramas.}
\usage{
plot_histograma(valores, estacoes, tipo = "Q", colunas = 10)
}
\arguments{
\item{valores}{Vetor numérico com as leituras a serem plotadas.}

\item{estacoes}{Vetor com as estações correspondentes a cada uma das
leituras.}

\item{tipo}{Variável que corresponde ao tipo de leitura, as opções são:
\code{Q} ou \code{P}, para vazões ou precipitações, respectivamente.}

\item{colunas}{Número de colunas a serem plotadas.}

\item{datas}{Vetor com as datas das leituras.}
}
\description{
Funções de plotagem de histogramas.
}
\details{
O histograma gerado separa por cores os dados por estação. Os
  gráficos são feitos com auxílio do pacote \code{\link[ggplot2]{ggplot2-package}}.
}
