% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histograma}
\alias{plot_histograma}
\title{Funções de plotagem de histogramas.}
\usage{
plot_histograma(dados, col_valores, colunas = 10)
}
\arguments{
\item{dados}{Data frame com a série hidrológica a ser utilizada.}

\item{col_valores}{O nome da coluna com os valores a serem utilizados.}

\item{colunas}{Número de colunas a serem plotadas.}
}
\description{
Funções de plotagem de histogramas.
}
\details{
O histograma gerado separa por cores os dados por estação. Os
  gráficos são feitos com auxílio do pacote \code{\link[ggplot2]{ggplot2-package}}.
}
\examples{
# Histograma da estação de Fluviópolis:
plot_histograma(fluviopolis, col_valores = "Q", colunas = 10)

}
