% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qx.R
\name{qx}
\alias{qx}
\title{Função de vazões médias em dias consecutivos.}
\usage{
qx(dados, x = 7)
}
\arguments{
\item{dados}{Data frame obtida após o uso da função \code{\link{organizar}}}

\item{x}{Número de dias consecutivos. O valor padrão é 7, correspondendo a
série de vazões Q7.}
}
\description{
Calcula a média das vazões a cada \code{x} dias consecutivos.
}
\details{
A função retorna uma data frame com as colunas de \code{Est},
  \code{Data} e \code{Qx}. A média só será realizada caso existam \code{x}
  dias consecutivos.
}
\examples{
# Cálculo das Q7:
qx(fluviopolis)

# Cálculo de Q15:
qx(fluviopolis, x = 15)

}
